refactor: decompose neural_network.py into architecture, trainer, and script exporter
Break the 1354-line monolithic NeuralNetworkInterface into three
focused modules:

- nn_architecture.py: Enums and dataclasses
- nn_trainer.py: NeuralNetworkTrainer with training logic
- nn_script_exporter.py: Script generation for PyTorch/TF/sklearn

The original neural_network.py is now a thin delegation facade
preserving full backward compatibility. Add 38 comprehensive tests.
